+++ src/main/java/com/solidarize/repository/UserRepository.java
@@ -1,5 +1,7 @@
 package com.solidarize.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,6 +10,6 @@ import com.solidarize.model.User;
 
 @Repository
 public interface UserRepository extends CrudRepository<User, Integer> {
-	User findById(Integer id);
+	User findById(BigInteger id);
 	User findByLogin(String Login);
 }
